Apparently tools are no longer in beta
Leaving this in actually broke this header:
    anthropic-beta: max-tokens-3-5-sonnet-2024-07-15

I think what's going on is that the server's HTTP implementation doesn't
support duplicate HTTP headers, but the HTTP spec does. It's likely that
they use a String => String mapping for headers, which isn't correct but
is surprisingly common. Unfortunately, the implication is that you can
only enable one beta feature at a time.